its working. Celebrate!

+++ admin_portal.php
@@ -1,6 +1,7 @@
 <?php
 
     include_once __DIR__ . '/model/product.php';
+    include_once __DIR__ . '/model/color.php';
     include_once __DIR__ . '/include/functions.php';
 
     // Set up configuration file and create database
@@ -14,18 +15,36 @@
     try 
     {
         $productDatabase = new Product($configFile);
+        $colorDatabase = new Color($configFile);
     } 
     catch ( Exception $error ) 
     {
         echo "<h2>" . $error->getMessage() . "</h2>";
     }   
     // If POST, delete the requested car before listing all
-    if (postRequest()) {
-        $product_id = filter_input(INPUT_POST, 'productID');
-        $productDatabase->deleteProduct($product_id);
+    if(postRequest()){
+        if(isset($_POST['delete'])){
+            $product_id = filter_input(INPUT_POST, 'productID');
+            
+            $productDatabase->deleteProduct($product_id);
 
+            $productDatabase->getProduct();
+            
+        }
+    }else{
+        //$productList = $productDatabase->getProduct();
+        echo "No Post Request";
     }
     $productList = $productDatabase->getProduct();
+    // if (postRequest()) {
+    //     $product_id = filter_input(INPUT_POST, 'productID');
+    //     $productDatabase->deleteProduct($product_id);
+    //     $productDatabase->getProduct();
+
+    // }
+
+
+
     
 ?>
 
@@ -68,17 +87,38 @@
 
         <tr>
             <td> 
-                <form action="admin_portal.php" method="post">
-            </td>                       
-            <td class="col-data" style="display: none;"><?php echo $row['productID']; ?></td>
-            <td class="col-data"><a href="update.php?action=Update&productID=<?php echo $row['productID']; ?>" style="color: blue;"><?php echo $row['productName']; ?></a></td>
-            <td class="col-data"><?php echo $row['categoryID']; ?></td>
-            <td class="col-data"><?php echo $row['colorID'];?></td>
+                </td>                       
+                <td class="col-data" style="display: none;" name="productID"><?php //echo $row['productID']; ?></td>
+                <td class="col-data"><a href="update.php?action=Update&productID=<?php echo $row['productID']; ?>" style="color: blue;"><?php echo $row['productName']; ?></a></td>
+                <td class="col-data"><?php     
+            if($row['categoryID'] == 1){
+                $category_desc = "shirt";
+            }
+            elseif($row['categoryID'] == 2){
+                $category_desc = "hoodie";
+            }
+            elseif($row['categoryID'] == 3){
+                $category_desc = "socks";
+            }
+            else{
+                $category_desc = "";
+            }
+            echo $category_desc; ?></td>
+            <td class="col-data"><?php     
+                $colorList = $colorDatabase->getColor($row['colorID']);
+                foreach($colorList as $colorRow):
+                    $colorDesc = $colorRow['colorDesc'];
+                    echo $colorDesc;
+                endforeach;    
+                ?></td>
             <td class="col-data"><?php echo $row['productSize']; ?></td>
             <td class="col-data"><?php echo $row['productQuantity']; ?></td>
-            <td class="col_data"><i class="fa-solid fa-trash-can"></i></td>
-        </form><!--end post form-->   
-                
+            <td class="col_data">
+                <form action="admin_portal.php" method="post">
+                    <input type="hidden" name="productID" value="<?php echo $row['productID']; ?>">
+                    <button class="fa-solid fa-trash-can" type="submit" name="delete"></button>
+                    </form><!--end post form-->   
+            </td>
         </tr>  
         <?php endforeach; ?>
     </table>
